Move parsing logic to module but it's not working...
